Move CloneResult to LibChorus
(cherry picked from commit 91a0e23f4b8a0bb6104c0cad0a8ef630b846dc5c)
