+++ src/LibChorus/LibChorus.csproj
@@ -391,6 +391,7 @@
     <Compile Include="Utilities\FilteredCatalog.cs" />
     <Compile Include="FileTypeHandlers\ChorusFileTypeHandlerCollection.cs" />
     <Compile Include="notes\EmbeddedMessageContentHandlerRepository.cs" />
+    <Compile Include="CloneResult.cs" />
   </ItemGroup>
   <ItemGroup>
     <BootstrapperPackage Include="Microsoft.Net.Framework.2.0">
